requirements.txt
playwright>=1.49.0
openai>=1.12.0
python-dotenv>=1.0.0
Pillow>=10.0.0
imagehash>=4.3.1
google-generativeai>=0.8.0
groq>=0.33.0
pyclean>=2.7.0
